Moved all routes into 'routes' directory for organization. Modified all files accessing new changes
