+++ .gitignore
@@ -61,8 +61,9 @@ model/user-data.csv
 backend/firebase_priv_key/munch-37564-firebase-adminsdk-ilv8p-e44bbd788a.json
 backend/.env
 backend/__pycache__
-backend/yelpAPI/__pycache__/
-backend/unit_tests/__pycache__/
+backend/routes/__pycache__
+backend/routes/yelpApi/__pycache__
+backend/unit_tests/__pycache__
 backend/.coverage
 backend/firestore-export
 backend/firebase-debug.log
